move & simplify create_table_multi_index
